==> Code/01.Fundamentals.ipynb <==
### Recommendation

**A. Content Based** : Recommend Product or Content based on **Similar Attributes**.
- I Saw a Comedy Movie Hera Pheri and Rated 5 Star
- I Saw a Comedy Movie Dhamaal and Rated 5 Star
- YouTube Started Recommending me more Comedy Movies ( Content )  

**Advantage** : Works even when Product or Content has no Rating but Attribute is Important for every Product or Content.

**B. Collaborative Filtering** : Recommend Product and Content based on Past User **Ratings** not based on Attributes.
- I Saw Avengers and Rated 5 Star | My Friend Akash Saw Avengers and Rated 4 Star
- I Saw Underground and Rated 4 Star | Akash Saw Underground and Rated 5 Star
- I Saw Captain America and Rated 5 Star | Hotstar Recommended Captain America to Akash.

**Advantage** : No need of Attributes, but User Reviews are very Important.

**Real Life Examples**

1. Social Media 
- Facebook : Posts and Friends.
- Pinterest : Products, Posts and Pins.
- Instagram : Posts, Timeline and Friends.

2. Online Content : YouTube, Hotstar and Netflix ( Shows, Movies and Contents )

3. Music Service : YouTube Music, Spotify ( Playlist ) and Apple Music Songs.

4. Ecommerce : Amazon, Flipkart and Myntra Products.

5. Banking and Insurance : Best Policy, Loan Amount and Insurance.

Recommender System ( Takes User Past Experience )
- Predicts the **Rating** the User would give the Product.
- Recommend the Products which will be liked by the User.
- Rank Products by User Preference.
- Product Similarity ( If User Buy Mobile he will also Buy Case )
- Find Similar User Preferences. ( Peoples mostly Buy this also )

If the User is New, He / She cannot be Recommended unless any Rating is Submitted.
- Until then Products Suggested are Based on his Search Interests and Similar Products.

import numpy as np
import pandas as pd
import matrix_factorization_utilities

ratings = [5,2,1,2,3,4,5,3,2,1]
scale10 = [i*2 for i in ratings]
print(f'Rating on the Scale of 10 : {scale10}')

Numpy is Faster and Efficient than List for Vector and Matrix Operations.

ratings = np.array([5,2,1,2,3,4,5,3,2,1])
scale10 = ratings * 2
print(f'Rating on the Scale of 10 : {scale10}')

ratings = pd.read_csv('../Data/Ratings.csv')
ratings.head()

movies = pd.read_csv('../Data/Movies.csv')
movies.head()

aggfunc = np.max : Replace Duplicates by Max Rating ( User may have Rated Same Movie Twice )

Creates a Sparse Matrix

rating = pd.pivot_table(ratings, index='user_id', columns='movie_id', aggfunc=np.max)
rating.head()

rating.to_csv('../Data/Rating_Matrix.csv', index=False, na_rep='')

### Building a Recommendation System

1. Create a Matrix of known User Reviews
2. Factor out a U Matrix and M Matrix from known Reviews.
3. Multiply the U and M Matrices we found to get Review Scores for every User and every Movie.

### Matrix Factorization

U x M = Movie Ratings

U : User Attributes

M : Movie Attributes

U, M = matrix_factorization_utilities.low_rank_matrix_factorization(rating.values, 
                                                          num_features = 15, 
                                                          regularization_amount = 0.1)

predicted_rating = np.matmul(U, M)

predicted_rating = pd.DataFrame(data=predicted_rating, 
                                 index=rating.index, 
                                 columns=rating.columns)
predicted_rating.to_csv('../Data/Predicted_Rating.csv')

predicted_rating.head()

M = np.transpose(M)

Get Movie Name and Genre

movie_info = movies.loc[5]

print(f'We are Finding Movies Similar to this Movie')
print(f'Movie Title : {movie_info.title}')
print(f'Movie Genre : {movie_info.genre}')

Feature's for Movie 1 founded by Matrix Factorization

movie_feature = M[5 - 1]
print(f'Attributes of the Movie \n{movie_feature}')

Main Logic to Find Similar Movie

Subtract Current Movie Feature from every other Movie

difference = M - movie_feature

Take Absolute Value of Difference ( All Members to be Positive )

absolute_difference = np.abs(difference)

Each Movie has 15 Features

total_difference = np.sum(absolute_difference, axis = 1)

Create a New Column in the Movie List with the Difference Score for each Movie.

movies['difference_score'] = total_difference

Sort the Movie List by Score Difference

sorted_movie = movies.sort_values('difference_score')

print(f'The Five Most Similar Movies are :')
print(sorted_movie[['title', 'difference_score']].head())
