==> Code/03.Movie Recommendation.ipynb <==
import numpy as np
import pandas as pd
import pickle

Import Movies Data Set

movies = pd.read_csv('../Data/Movies.csv', index_col='movie_id')

Load Predictions and Features 

U = pickle.load(open('../Data/User_Rating_Features.dat', mode='rb'))
M = pickle.load(open('../Data/Movie_Features.dat', mode='rb'))
predictions = pickle.load(open('../Data/Predicted_Ratings.dat', mode='rb'))

Recommendation

user_id = int(input('Enter User ID to get Recommendations between 1 and 100 :'))
print(f'\nMovies Recommended :')

user_ratings = predictions[user_id - 1]

movies['rating'] = user_ratings
movies = movies.sort_values(by=['rating'], ascending=False)

print(f'{movies[["title", "genre", "rating"]].head()}')
