==> Code/04.Handle New User.ipynb <==
### Handling First Time Users

1. Do not make any Recommendations.

2. Recommend based on Product Similarity.

3. Recommend based on overall Average Ratings.

import numpy as np
import pandas as pd
import pickle
import matrix_factorization_utilities

Import Ratings Data Set

ratings = pd.read_csv('../Data/Ratings.csv')
ratings.head(3)

Convert User Ratings into Matrix

matrix = pd.pivot_table(ratings, 
                        index='user_id', 
                        columns='movie_id', 
                        aggfunc=np.max)

Normalize Ratings ( Center them around the Mean )

normalized_ratings, mean = matrix_factorization_utilities.normalize_ratings(matrix.values)

Apply Matrix Factorization to find the Latent Features ( Hidden Features )

U, M = matrix_factorization_utilities.low_rank_matrix_factorization(normalized_ratings, 
                                                                    num_features=15, 
                                                                    regularization_amount=1.1)

Predicted Ratings : U x M

predicted_ratings = np.matmul(U, M)

Adding back Mean Ratings 

predicted_ratings += mean

Save User Rating Features, Movie Features and Predicted Ratings Features 

pickle.dump(U, open('../Data/New_User_Rating_Features.dat', mode='wb'))
pickle.dump(M, open('../Data/New_User_Movie_Features.dat', mode='wb'))
pickle.dump(predicted_ratings, open('../Data/New_User_Predicted_Ratings.dat', mode='wb'))
pickle.dump(mean, open('../Data/New_User_Mean.dat', mode='wb'))
