==> Code/05.New User Recommendation.ipynb <==
### Library 

import numpy as np
import pandas as pd
import pickle

Import Mean Data File

mean = pickle.load(open('../Data/New_User_Mean.dat', mode='rb'))

Import Movie Data Set

movies = pd.read_csv('../Data/Movies.csv', index_col='movie_id')
movies.head(3)

Average User Ratings

user_ratings = mean
print('Movie Recommended :')
movies['rating'] = user_ratings
movies = movies.sort_values(by = ['rating'], ascending = False)
print(f'{movies[["title", "genre", "rating"]].head()}')
