==> Code/06.Finding Similar Movie.ipynb <==
import numpy as np
import pandas as pd
import pickle

Import Movie Prediction

M = pickle.load(open('../Data/New_User_Movie_Features.dat', mode='rb'))
M = np.transpose(M)

Import Movies Data Set

movies = pd.read_csv('../Data/Movies.csv', index_col='movie_id')

Choose a Movie to Find Similar Movies

movie_info = movies.loc[5]
print(f'We are Finding Movies Similar to this Movie :')
print(f'Movie Title : {movie_info.title}')
print(f'Genre : {movie_info.genre}')

Get Features for Movie 1 found via Matrix Factorization

current_movie_features = M[5-1]
print('The Attributes of the Movie are :')
print(f'{current_movie_features}')

Main Logic behind Finding Similar Movies

1. Subtract the Current Movie's Features from every other Movie's Features

difference = M - current_movie_features

2. Take the Absolute Value of that Difference ( All Numbers are Positive )

absolute_difference = np.abs(difference)

3. Total Difference Score 

total_difference = np.sum(absolute_difference, axis=1)

4. Create a New Column in Movie List with Difference Score for Each Movie.

movies['difference_score'] = total_difference

5. Sort the Movie List by Difference Score, from Least Difference to Most Difference.

sorted_movie_list = movies.sort_values('difference_score')

Similar Movies

print(f'The 5 Most Similar Movies are:')
print(sorted_movie_list[['title', 'difference_score']].head())
